1.用户查询报名比赛接口使用了用户查询成绩接口，并且对比赛进行了按开始时间从早到晚排序 2.取消报名接口进行是否为报名期间判断 3.报名接口进行是否为报名期间判断

+++ sport-game/src/main/java/com/ruoyi/system/controller/SportRegistrationsController.java
@@ -1,5 +1,6 @@
 package com.ruoyi.system.controller;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
@@ -152,9 +153,19 @@ public class SportRegistrationsController extends BaseController
     public TableDataInfo userRegisterationslist(SportRegistrations sportRegistrations)
     {
         sportRegistrations.setUserId(SecurityUtils.getUserId());
+        //查询当前用户报名的比赛集合并且分页
         startPage();
-        List<SportRegistrations> sportRegistrationsList= sportRegistrationsService.userRegisterationslist(sportRegistrations);
-        return getDataTable(sportRegistrationsList);
+        List<SportRegistrations> sportRegistrationsList= sportRegistrationsService.selectSportRegistrationsList(sportRegistrations);
+
+        //遍历集合获取用户每项比赛成绩
+        List<UserSportGradeVo> userSportGradeVoList=sportRegistrationsList.stream().map((item)->{
+            UserSportGradeVo userSportGradeVo= sportRegistrationsService.selectUserSportGrade(item.getGameId());
+            return userSportGradeVo;
+        }).collect(Collectors.toList());
+
+        //对比赛成绩集合进行按比赛时间从早到晚排序
+        Collections.sort(userSportGradeVoList);
+        return getDataTable(userSportGradeVoList);
     }
 
     /**
@@ -164,58 +175,7 @@ public class SportRegistrationsController extends BaseController
     @GetMapping("/user/{gameId}")
     public AjaxResult userScore(@PathVariable("gameId") Long gameId)
     {
-        String redisKey = "userSportGrade:" + SecurityUtils.getUserId()+":"+gameId;
-        UserSportGradeVo cacheObject = (UserSportGradeVo)redisCache.getCacheObject(redisKey);
-        if(!ObjectUtils.isEmpty(cacheObject)){
-            return AjaxResult.success(cacheObject);
-        }
-        SportRegistrations sportRegistrations=new SportRegistrations();
-        sportRegistrations.setGameId(gameId);
-        //查询参加该比赛的所有报名集合
-        List<SportRegistrations> sportRegistrationsList = sportRegistrationsService.selectSportRegistrationsList(sportRegistrations);
-
-        SportRegistrations userSportRegistrations=new SportRegistrations();//用于从集合中获得用户的报名信息
-        //遍历集合从中获取当前用户成绩
-        for(SportRegistrations temp:sportRegistrationsList){
-            if(SecurityUtils.getUserId()==temp.getUserId()){
-                userSportRegistrations=temp;
-            }
-        }
-        if(userSportRegistrations.getScore()==null){
-            return AjaxResult.error("未查询到用户成绩");
-        }
-        UserSportGradeVo userSportGradeVo=new UserSportGradeVo();
-
-        //根据比赛id查出比赛的item_id
-        SportGames sportGames = sportGamesService.selectSportGamesById(gameId);
-        Long itemId = sportGames.getItemId();
-        //根据item_id查出降序还是升序
-        SportItem sportItem = sportItemService.selectSportItemById(itemId);
-        Long isDesc = sportItem.getIsDesc();
-
-        sportGames.setItem(sportItem);
-        userSportRegistrations.setGame(sportGames);
-        userSportGradeVo.setSportRegistrations(userSportRegistrations);
-
-        //根据升序降序对用户进行排名
-        Long userOrder=1L;
-        for(SportRegistrations temp:sportRegistrationsList){
-           if(temp.getScore()!=null){
-               if(isDesc==1){//降序
-                   if(temp.getScore()!=null&&userSportRegistrations.getScore()<temp.getScore()){
-                       userOrder++;
-                   }
-               }else {//升序
-                   if(temp.getScore()!=null&&userSportRegistrations.getScore()>temp.getScore()){
-                       userOrder++;
-                   }
-               }
-           }
-        }
-        userSportGradeVo.setUserOrder(userOrder);
-        //将用户当前成绩缓存，缓存时长设置为1h
-        redisCache.setCacheObject(redisKey,userSportGradeVo,1, TimeUnit.HOURS);
-
+        UserSportGradeVo userSportGradeVo = sportRegistrationsService.selectUserSportGrade(gameId);
         return AjaxResult.success(userSportGradeVo);
     }
 
@@ -242,6 +202,11 @@ public class SportRegistrationsController extends BaseController
             return AjaxResult.error("用户性别错误");
         }
 
+        //判断报名时是否在报名时间段内
+        if(sportGames.getStatus()!=0){
+            return AjaxResult.error("未在报名时间段内");
+        }
+
         sportRegistrations.setUserId(SecurityUtils.getUserId());
         sportRegistrations.setUpdateTime(DateUtils.getNowDate());
         sportRegistrations.setCreateTime(DateUtils.getNowDate());
@@ -256,6 +221,13 @@ public class SportRegistrationsController extends BaseController
     @DeleteMapping("/user/{gameId}")
     public AjaxResult deleteUserRegistrations(@PathVariable("gameId") Long gameId)
     {
+        SportGames sportGames = sportGamesService.selectSportGamesById(gameId);
+
+        //判断取消报名时是否在报名时间段内
+        if(sportGames.getStatus()!=0){
+            return AjaxResult.error("未在报名时间段内");
+        }
+
         return toAjax(sportRegistrationsService.deleteUserRegistrations(SecurityUtils.getUserId(),gameId));
     }
